moved serialization funcs from App to Serializer
