+++ src/application.ts
@@ -17,6 +17,7 @@ import unpick from "./utils/unpick";
 
 import ApplicationInstance from "./application-instance";
 import JsonApiSerializer from "./serializers/serializer";
+import Password from "./attribute-types/password";
 
 export default class Application {
   namespace: string;
@@ -108,7 +109,8 @@ export default class Application {
           op.data.relationships.hasOwnProperty(relName)
       )
       .reduce((relationAttributes, relName) => {
-        const key = schemaRelationships[relName].foreignKeyName || this.serializer.relationshipToColumn(relName, primaryKey);
+        const key =
+          schemaRelationships[relName].foreignKeyName || this.serializer.relationshipToColumn(relName, primaryKey);
         const value = (<ResourceRelationshipData>op.data.relationships[relName].data).id;
 
         return {
@@ -165,68 +167,7 @@ export default class Application {
 
     const dataArrayed = Array.isArray(data) ? data : [data];
     const resource = await this.resourceFor(dataArrayed[0].type);
-    return dataArrayed.map(record => this.serializeResource(record, resource));
-  }
-
-  serializeResource(data: Resource, resource: typeof Resource): Resource {
-    const resourceSchema = resource.schema;
-    const schemaRelationships = resourceSchema.relationships;
-    const relationshipsFound = Object.keys(schemaRelationships)
-      .filter(relName => schemaRelationships[relName].belongsTo)
-      .filter(
-        relName =>
-          data.attributes.hasOwnProperty(schemaRelationships[relName].foreignKeyName) ||
-          data.attributes.hasOwnProperty(`${relName}Id`)
-      )
-      .map(relName => ({
-        name: relName,
-        key: schemaRelationships[relName].foreignKeyName || `${relName}Id`
-      }));
-
-    data.relationships = relationshipsFound.reduce(
-      (relationships, relationship) => ({
-        ...relationships,
-        [relationship.name]: {
-          id: data.attributes[relationship.key],
-          type: schemaRelationships[relationship.name].type().type
-        }
-      }),
-      data.relationships as any
-    );
-
-    data.attributes = unpick(
-      data.attributes,
-      relationshipsFound
-        .map(relationship => relationship.key)
-        .filter(relationshipKey => !Object.keys(resourceSchema.attributes).includes(relationshipKey))
-    );
-    Object.keys(data.relationships).filter(relName => data.relationships[relName]).forEach(relName => {
-      const fkName = schemaRelationships[relName].belongsTo
-        ? "id"
-        : schemaRelationships[relName].type().schema.primaryKeyName || DEFAULT_PRIMARY_KEY;
-
-      const relationship = this.serializeRelationship(
-        (data.relationships[relName] as unknown) as Resource | Resource[],
-        fkName
-      );
-      data.relationships[relName] = {
-        data: relationship
-      };
-    });
-
-    return data;
-  }
-
-  serializeRelationship(relationships: Resource | Resource[], primaryKeyName: string) {
-    if (Array.isArray(relationships)) {
-      return relationships.map(relationship => this.serializeRelationship(relationship, primaryKeyName));
-    }
-    relationships.id = relationships[primaryKeyName || DEFAULT_PRIMARY_KEY];
-    if (!relationships.id) {
-      return null;
-    }
-
-    return pick(relationships, ["id", "type"]);
+    return dataArrayed.map(record => this.serializer.serializeResource(record, resource));
   }
 
   // TODO: remove type any for data.relationships[relationshipName]
@@ -252,10 +193,15 @@ export default class Application {
 
             if (resource[pkName]) {
               includedData.push(
-                this.serializeResource(
+                this.serializer.serializeResource(
                   new relatedResourceClass({
                     id: resource[pkName],
-                    attributes: unpick(resource, [pkName])
+                    attributes: unpick(resource, [
+                      pkName,
+                      ...Object.keys(relatedResourceClass.schema.attributes).filter(
+                        attribute => relatedResourceClass.schema.attributes[attribute] === Password
+                      )
+                    ])
                   }),
                   relatedResourceClass
                 )
@@ -272,10 +218,15 @@ export default class Application {
 
           if (resource[pkName]) {
             includedData.push(
-              this.serializeResource(
+              this.serializer.serializeResource(
                 new relatedResourceClass({
                   id: resource[pkName],
-                  attributes: unpick(resource, [pkName])
+                  attributes: unpick(resource, [
+                    pkName,
+                    ...Object.keys(relatedResourceClass.schema.attributes).filter(
+                      attribute => relatedResourceClass.schema.attributes[attribute] === Password
+                    )
+                  ])
                 }),
                 relatedResourceClass
               )
